rest_api_service/pyproject.toml
[tool.poetry]
name = "rest-api-server"
version = "0.1.0"
description = ""
authors = ["Yunho Jang, yunho0520@gmail.com"]

[tool.poetry.dependencies]
python = "~3.7"
Flask = "~1.1"
Flask-Log-Request-ID = "~0.10.1"
gunicorn = "20.0.4"
pymongo = "3.10.1"
pymodm = "~0.4.3"
fastjsonschema = "~2.14"
nginx-config-builder = "1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-mongodb = "2.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
